+++ home-application/src/main/java/com/softserveinc/ita/homeproject/application/api/FileApiImpl.java
@@ -4,7 +4,7 @@ import javax.annotation.security.PermitAll;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.Provider;
 
-import com.softserveinc.ita.homeproject.homeservice.service.cooperation.data_damping.ImportExportTemplateFileService;
+import com.softserveinc.ita.homeproject.homeservice.service.cooperation.data.transfer.ImportExportTemplateFileService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
