+++ home-service/src/main/java/com/softserveinc/ita/homeproject/homeservice/service/cooperation/data/transfer/ImportExportTemplateFileServiceImpl.java
@@ -1,4 +1,4 @@
-package com.softserveinc.ita.homeproject.homeservice.service.cooperation.data_damping;
+package com.softserveinc.ita.homeproject.homeservice.service.cooperation.data.transfer;
 
 import java.io.File;
